requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
librosa>=0.10.0
yt-dlp>=2023.7.6
Pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.20.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
spotipy>=2.23.0
mutagen>=1.47.0
safetensors>=0.4.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0

